Update YAML and Database
- Updated database (locally) to include a primary key that is  auto incrementing. Will most likely host database online.
- Updated YAML structure.

+++ src/main/java/org/sapd/sapdhub/ArrestRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface ArrestRepository extends JpaRepository<Arrest, BigInteger>, JpaSpecificationExecutor<Arrest> {
+public interface ArrestRepository extends JpaRepository<Arrest, Integer>, JpaSpecificationExecutor<Arrest> {
 
     //void deleteByReport_id(BigInteger reportId);
     //Optional<Arrest> findById(BigInteger id);
